reorg bindings/wasm to remove wasm workarounds
