+++ crates/xmtp_api_d14n/src/queries/client_bundle.rs
@@ -23,6 +23,7 @@ pub enum ClientKind {
     V3,
     Hybrid,
 }
+
 impl std::fmt::Display for ClientKind {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         use ClientKind::*;
